Integral package has been moved from stats to math. Derivative package has been created with a class implementing numerical differentiation.
