+++ src/repicea/math/integral/AbstractGaussQuadrature.java
@@ -16,7 +16,7 @@
  *
  * Please see the license at http://www.gnu.org/copyleft/lesser.html.
  */
-package repicea.stats.integral;
+package repicea.math.integral;
 
 import java.util.ArrayList;
 import java.util.Collections;
