+++ src/repicea/math/integral/AbstractNumericalIntegrationMethod.java
@@ -16,7 +16,7 @@
  *
  * Please see the license at http://www.gnu.org/copyleft/lesser.html.
  */
-package repicea.stats.integral;
+package repicea.math.integral;
 
 import java.io.Serializable;
 import java.util.ArrayList;
