+++ src/repicea/stats/model/glm/measerr/GLMNormalClassicalMeasErrorDefinition.java
@@ -29,12 +29,12 @@ import repicea.math.MathematicalFunction;
 import repicea.math.Matrix;
 import repicea.math.ProductFunctionWrapper;
 import repicea.math.SymmetricMatrix;
+import repicea.math.integral.TrapezoidalRule;
 import repicea.stats.data.DataSet;
 import repicea.stats.data.GenericStatisticalDataStructure;
 import repicea.stats.data.StatisticalDataStructure;
 import repicea.stats.distributions.utility.GaussianUtility;
 import repicea.stats.functions.EmpiricalDistributionProbabilityDensityFunction;
-import repicea.stats.integral.TrapezoidalRule;
 import repicea.stats.model.CompositeLogLikelihoodWithExplanatoryVariables;
 import repicea.stats.model.IndividualLogLikelihood;
 import repicea.stats.model.Likelihood;
